Consolidate playlistFavorites routes tests into one file
